requirements.txt
PyQt6>=6.6.0
pyserial>=3.5
pymodbus>=3.5.0
